requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blessed==1.20.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
crawlee==0.3.5
curl_cffi==0.7.1
dnspython==2.6.1
docutils==0.21.2
editor==1.6.6
email_validator==2.2.0
eval_type_backport==0.2.0
fastapi==0.114.0
fastapi-cli==0.0.5
filelock==3.16.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
inquirer==3.4.0
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
playwright==1.47.0
psutil==6.0.0
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pyee==12.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.2
readchar==4.2.0
requests==2.32.3
requests-file==2.1.0
rich==13.8.0
runs==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.38.5
text-unidecode==1.3
tldextract==5.1.2
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
urllib3==2.2.3
useragent==0.1.1
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.0.1
xmod==1.8.1
